Rearranged sections and TODO list, and cleared old TODO items
Reordered the "Commands" chapter, at the same time as moving the
introductory paragraph to the "Arguments" chapter and its "Variables"
section to this chapter.

The remaining sections in the "Arguments" chapter were all on
expansion, globbing, and predicates (collectively referred to by the
Zsh manual, after which this behaviour was modelled, as "filename
generation"), which were moved into their own chapter, now called
"Expansion".